% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{fetch_open_meteo}
\alias{fetch_open_meteo}
\title{Fetch Open-Meteo Data}
\usage{
fetch_open_meteo(lat, lon, start_date, end_date = as.character(Sys.Date()))
}
\arguments{
\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{start_date}{Start date, "YYYY-MM-DD".}

\item{end_date}{End date, "YYYY-MM-DD". Defaults to today.}
}
\value{
A data frame with Date, tmax, tmin (Fahrenheit).
}
\description{
Fetch Open-Meteo Data
}
