% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenology.R
\name{plot_phenology_trend}
\alias{plot_phenology_trend}
\title{Plot Cumulative Degree-Days against Trap Counts}
\usage{
plot_phenology_trend(df, save_plot = FALSE)
}
\arguments{
\item{df}{The data frame outputted by \code{calc_pest_phenology}.}

\item{save_plot}{Logical. If TRUE, saves the plot to your working directory.}
}
\value{
A ggplot object.
}
\description{
Plot Cumulative Degree-Days against Trap Counts
}
