% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackTrap-package.R
\docType{package}
\name{TrackTrap-package}
\alias{TrackTrap}
\alias{TrackTrap-package}
\title{TrackTrap: Model Cumulative Growing Degree-Days for Pest Monitoring}
\description{
Pest monitoring is crucial, especially during the early season, to understand the distribution and the proliferation of the target pest. Raw count data from pest monitoring/traps can be coupled with environmental variables such as temperature, growing degree-day ('GDD') etc. to get useful insights about the pest phenology. This package pulls temperature data from the 'Daymet' application programming interface ('API', \url{https://daymet.ornl.gov}), or 'Open-Meteo' ('API', \url{https://open-meteo.com/}) or manually from the California Irrigation Management Information System ('CIMIS', \url{https://cimis.water.ca.gov}) as a csv import, for a user-specified time period and calculates cumulative growing degree-days. Users provide pest development thresholds (lower and upper temperatures), pest of concern, and the geographic coordinates of the trap location to track emergence and phenology.
}
\author{
\strong{Maintainer}: Santosh Bhandari \email{santoshbhandari4556@gmail.com}

Authors:
\itemize{
  \item Santosh Bhandari \email{santoshbhandari4556@gmail.com}
  \item Lalit Kharel \email{klalitkharel@gmail.com}
  \item Mahesh Ghimire \email{maesghimire@gmail.com}
}

}
\keyword{internal}
